Android: ro'yxatdan o'tkazish oqimi E1-E6
Har bir qadam tugmani emas, IMKONIYATNING O'ZINI tekshiradi.
RequestPermission natijasi ataylab e'tiborsiz qoldiriladi — qatorni
CapabilityChecks.check() hal qiladi: bir soniyalik AudioRecord yozuvi
xotirada tekshiriladi (butunlay nol bo'lsa = OEM bloklagan, chunki jim xona
raqamli sukunat emas), call log bitta qator so'raladi, ACTION_CALL
hal qilinadimi, batareya istisnosi bormi.

Natija 'ruxsat berilgan, lekin ishlamayapti' bo'lsa — alohida jumla:
dialog buni tuzatmaydi, sozlamalarga o'tish kerak. Har bir natija
serverga darhol yuboriladi, ya'ni panel to'siqni ikki daqiqada ko'radi.

E1 da 'tayyor' so'zi umuman yo'q: kodni almashtirish faqat vaqtinchalik
token beradi.

N40 o'lchanadi: monotonik soatda qadam vaqti, va har bir ekranda doimiy
'Yordam kerak' tugmasi — qotib qolgan ro'yxatdan o'tish sukunat emas,
hodisa bo'lishi kerak.

+++ android/app/src/main/kotlin/uz/bonvi/call/data/remote/dto/DeviceCallIn.kt
@@ -13,6 +13,11 @@
     "UnusedImport"
 )
 
+// PATCHED by android/scripts/widen_int64.py: the contract declares this
+// field as a formatless integer, which generates as a 32-bit kotlin.Int
+// and cannot hold its real range. Delete the patch once the server emits
+// format: int64 — Int64WireContractTest says when.
+
 package uz.bonvi.call.data.remote.dto
 
 import uz.bonvi.call.data.remote.dto.AppVariant
@@ -62,7 +67,7 @@ data class DeviceCallIn (
 
     /* System.currentTimeMillis() at call start — raw evidence for skew. */
     @Json(name = "device_epoch_ms")
-    val deviceEpochMs: kotlin.Int,
+    val deviceEpochMs: kotlin.Long,
 
     /* IANA name. */
     @Json(name = "device_timezone")
